airflow-docker/requirements.txt
apache-airflow==2.7.3

psycopg2-binary==2.9.9
SQLAlchemy==1.4.49



pandas==2.2.2
requests==2.31.0
